Moved norm back from stats to linalg, which is more coherent with existing toolsets
